modified api documentation into different files and added comments including questions
